chore: domain/project -> project로 폴더 이동
