--- src/main/java/com/scg/stop/domain/project/domain/ProjectType.java
@@ -1,9 +1,0 @@
-package com.scg.stop.domain.project.domain;
-
-public enum ProjectType {
-
-    RESEARCH_AND_BUSINESS_FOUNDATION,
-    LAB,
-    STARTUP,
-    CLUB
-}
