--- src/main/java/com/scg/stop/domain/project/domain/Role.java
@@ -1,7 +1,0 @@
-package com.scg.stop.domain.project.domain;
-
-public enum Role {
-
-    STUDENT,
-    PROFESSOR
-}
