+++ src/main/java/com/scg/stop/project/domain/AwardStatus.java
@@ -1,0 +1,11 @@
+package com.scg.stop.project.domain;
+
+public enum AwardStatus {
+
+    NONE,
+    FIRST,  // 대상
+    SECOND, // 최우수상
+    THIRD,  // 우수상
+    FOURTH, // 장려상
+    FIFTH   // 인기상
+}
