+++ src/main/java/com/scg/stop/project/domain/ProjectCategory.java
@@ -1,0 +1,13 @@
+package com.scg.stop.project.domain;
+
+public enum ProjectCategory {
+
+    COMPUTER_VISION,
+    SYSTEM_NETWORK,
+    WEB_APPLICATION,
+    SECURITY_SOFTWARE_ENGINEERING,
+    NATURAL_LANGUAGE_PROCESSING,
+    BIG_DATA_ANALYSIS,
+    AI_MACHINE_LEARNING,
+    INTERACTION_AUGMENTED_REALITY
+}
