+++ src/main/java/com/scg/stop/project/domain/ProjectType.java
@@ -1,0 +1,9 @@
+package com.scg.stop.project.domain;
+
+public enum ProjectType {
+
+    RESEARCH_AND_BUSINESS_FOUNDATION,
+    LAB,
+    STARTUP,
+    CLUB
+}
