+++ src/main/java/com/scg/stop/project/domain/Role.java
@@ -1,0 +1,7 @@
+package com.scg.stop.project.domain;
+
+public enum Role {
+
+    STUDENT,
+    PROFESSOR
+}
