+++ src/main/java/com/scg/stop/project/dto/response/CommentResponse.java
@@ -1,0 +1,31 @@
+package com.scg.stop.project.dto.response;
+
+import com.scg.stop.project.domain.Comment;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+public class CommentResponse {
+    private Long id;
+    private Long projectId;
+    private String userName;
+    private Boolean isAnonymous;
+    private String content;
+    private LocalDateTime createdAt;
+    private LocalDateTime updatedAt;
+
+    public static CommentResponse of(Comment comment){
+        return new CommentResponse(
+                comment.getId(),
+                comment.getProject().getId(),
+                comment.getIsAnonymous() ? null : comment.getUser().getName(), // 익명이 아닐 경우에만 사용자 이름을 반환
+                comment.getIsAnonymous(),
+                comment.getContent(),
+                comment.getCreatedAt(),
+                comment.getUpdatedAt()
+        );
+    }
+}
