+++ src/main/java/com/scg/stop/project/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package com.scg.stop.project.repository;
+
+import com.scg.stop.project.domain.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
