+++ src/main/java/com/scg/stop/user/domain/User.java
@@ -5,11 +5,10 @@ import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
 
-import com.scg.stop.auth.domain.request.RegisterRequest;
-import com.scg.stop.domain.project.domain.Comment;
-import com.scg.stop.domain.project.domain.FavoriteProject;
-import com.scg.stop.domain.project.domain.Inquiry;
-import com.scg.stop.domain.project.domain.Likes;
+import com.scg.stop.project.domain.Comment;
+import com.scg.stop.project.domain.FavoriteProject;
+import com.scg.stop.project.domain.Inquiry;
+import com.scg.stop.project.domain.Likes;
 import com.scg.stop.domain.proposal.domain.Proposal;
 import com.scg.stop.domain.video.domain.FavoriteVideo;
 import com.scg.stop.domain.video.domain.UserQuiz;
