+++ src/test/java/com/scg/stop/domain/project/controller/ProjectControllerTest.java
@@ -17,18 +17,19 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.scg.stop.configuration.AbstractControllerTest;
 
-import com.scg.stop.domain.project.domain.AwardStatus;
-import com.scg.stop.domain.project.domain.ProjectCategory;
-import com.scg.stop.domain.project.domain.ProjectType;
-import com.scg.stop.domain.project.domain.Role;
-import com.scg.stop.domain.project.dto.request.CommentRequest;
-import com.scg.stop.domain.project.dto.request.MemberRequest;
-import com.scg.stop.domain.project.dto.request.ProjectRequest;
-import com.scg.stop.domain.project.dto.response.CommentResponse;
-import com.scg.stop.domain.project.dto.response.FileResponse;
-import com.scg.stop.domain.project.dto.response.ProjectDetailResponse;
-import com.scg.stop.domain.project.dto.response.ProjectResponse;
-import com.scg.stop.domain.project.service.ProjectService;
+import com.scg.stop.project.controller.ProjectController;
+import com.scg.stop.project.domain.AwardStatus;
+import com.scg.stop.project.domain.ProjectCategory;
+import com.scg.stop.project.domain.ProjectType;
+import com.scg.stop.project.domain.Role;
+import com.scg.stop.project.dto.request.CommentRequest;
+import com.scg.stop.project.dto.request.MemberRequest;
+import com.scg.stop.project.dto.request.ProjectRequest;
+import com.scg.stop.project.dto.response.CommentResponse;
+import com.scg.stop.project.dto.response.FileResponse;
+import com.scg.stop.project.dto.response.ProjectDetailResponse;
+import com.scg.stop.project.dto.response.ProjectResponse;
+import com.scg.stop.project.service.ProjectService;
 import com.scg.stop.user.domain.User;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.DisplayName;
